Added support for Rgba32 textures.

+++ FinModelUtility/Fin/src/math/BitLogic.cs
@@ -28,8 +28,8 @@ namespace fin.math {
     //    Function to extract k bits from p position
     //    and returns the extracted value as integer
     /// </summary>
-    public static int ExtractFromRight(ushort number, int offset, int count)
-     // => (((1 << count) - 1) & (number >> (offset - 1)));
-    => (number >> offset) & ((1 << count) - 1);
+    public static uint ExtractFromRight(uint number, int offset, int count)
+        // => (((1 << count) - 1) & (number >> (offset - 1)));
+      => (uint) ((number >> offset) & ((1 << count) - 1));
   }
 }
